update error handling, update tests
