+++ Frameworks/MetalPetal/MTIError.h
@@ -26,7 +26,6 @@ typedef NS_ERROR_ENUM(MTIErrorDomain, MTIError) {
     //Texture loading errors
     MTIErrorUnsupportedCVPixelBufferFormat = 2001,
     MTIErrorTextureDimensionsMismatch = 2002,
-    MTIErrorTextureLoaderOptionNotSupported = 2003,
     MTIErrorTextureLoaderFailedToCreateCGContext = 2004,
     MTIErrorTextureLoaderFailedToCreateCGImage = 2005,
 
@@ -34,11 +33,12 @@ typedef NS_ERROR_ENUM(MTIErrorDomain, MTIError) {
     MTIErrorUnsupportedImageCachePolicy = 3001,
     
     //Kernel errors
-    MTIErrorDataBufferSizeMismatch = 4001,
-    MTIErrorParameterDataTypeNotSupported = 4002,
+    MTIErrorParameterDataSizeMismatch = 4001,
+    MTIErrorUnsupportedParameterType = 4002,
     MTIErrorMPSKernelInputCountMismatch = 4003,
     MTIErrorMPSKernelNotSupported = 4004,
     MTIErrorTextureBindingFailed = 4005,
+    MTIErrorParameterDataTypeMismatch = 4006,
     
     //Render errors
     MTIErrorEmptyDrawable = 5001,
